handle take damage

+++ client/Assets/_Project/Combat/_Interface/IDamageable.cs
@@ -3,6 +3,6 @@ using UnityEngine;
 namespace TopdownRPG.Combat {
     public interface IDamageable {
         GameObject GameObject { get; }
-        void TakeDamage(float damage);
+        void TakeDamage(int damage);
     }
 }
